% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Create a new NAFO document based on a template}
\usage{
draft(report_type = c("SCR", "STACFIS", "SCS"), create_dir = FALSE,
  edit = FALSE, ...)
}
\arguments{
\item{report_type}{The type of report to start.}

\item{create_dir}{\code{TRUE} to create a new directory for the document.}

\item{edit}{\code{TRUE} to edit the template immediately.}

\item{...}{Other arguments to pass to \code{\link[rmarkdown:draft]{rmarkdown::draft()}}.}
}
\description{
Create a draft of an R Markdown NAFO document
}
\details{
This is a light wrapper around \code{\link[rmarkdown:draft]{rmarkdown::draft()}}.
Consult that function for further details.
}
\examples{
\dontrun{
NAFOdown::draft("SCR")
}
}
