% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_leaf.R
\name{F_linear}
\alias{F_linear}
\alias{F_sigmoid}
\alias{F_lower}
\alias{F_upper}
\title{Calculate fishing mortality under the PA leaf}
\usage{
F_linear(B, Blim, Btrigger, Ftarget, Fmin = 0)

F_sigmoid(B, Blim, Btrigger, Ftarget, x50, k = 1, Fmin = 0)

F_lower(B, Blim, Btrigger, Ftarget, k = 1, Fmin = 0)

F_upper(B, Blim, Btrigger, Ftarget, k = 1, Fmin = 0)
}
\arguments{
\item{B}{Biomass vector}

\item{Blim}{Biomass limit reference point}

\item{Btrigger}{Biomass trigger reference point}

\item{Ftarget}{Target fishing mortality}

\item{Fmin}{Minimum fishing mortality (default: 0)}

\item{x50}{(for \code{F_sigmoid}) Proportion of the biomass range at which F = 0.5 * Ftarget}

\item{k}{(for \code{F_sigmoid}) Shape parameter for the sigmoid function (default: 1)}
}
\value{
Numeric vector of fishing mortality values
}
\description{
These functions calculate fishing mortality based on biomass using different harvest control rule (HCR) approaches:
}
\details{
\itemize{
\item \code{F_linear()} implements a linear HCR
\item \code{F_sigmoid()} implements a sigmoid HCR with configurable inflection point (x50)
\item \code{F_lower()} sets x50 = 0.75 for the lower bound of the leaf as per Annex 4 in \href{https://www.nafo.int/Portals/0/PDFs/COM-SC/2023/com-scdoc23-03.pdf}{Com-SC RBMS WP 23-03 Rev 2}
\item \code{F_upper()} sets x50 = 0.25 for the upper bound of the leaf as per Annex 4 in \href{https://www.nafo.int/Portals/0/PDFs/COM-SC/2023/com-scdoc23-03.pdf}{Com-SC RBMS WP 23-03 Rev 2}
}
}
