% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\docType{data}
\name{nafo-vals}
\alias{nafo-vals}
\alias{.nafo_lwd}
\alias{.nafo_height}
\alias{.nafo_width}
\alias{.nafo_asp}
\title{Helper values for consistent plot settings}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
.nafo_lwd

.nafo_height

.nafo_width

.nafo_asp
}
\description{
\itemize{
\item \code{.nafo_lwd}: this value is to be used in geom_line to define line width (it is 0.75 pt converted to mm)
\item \code{.nafo_height}: this value is to be used to define plot height in inches (it is 6.4 cm converted to inches)
\item \code{.nafo_width}: this value is to be used to define plot width in inches (it is 11.5 cm converted to inches)
\item \code{.nafo_asp}: this value is to be used to define the aspect ratio of a plot
}
}
\keyword{datasets}
