% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{nafo_pal}
\alias{nafo_pal}
\title{NAFO Colour Palettes}
\usage{
nafo_pal(palette = "logo", n = length(.nafo_cols))
}
\arguments{
\item{palette}{Type of palette. Options are "logo", "blues", "greens", "blue2green", and "hot_cold".}

\item{n}{Number of colours to return.}
}
\value{
A function that generates the specified palette.
}
\description{
This function returns a colour palette function based on the specified type.
}
\examples{
nafo_pal("logo")
nafo_pal("blues", 10)
nafo_pal("hot_cold", 10)

}
