% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAFOdown-package.R
\docType{package}
\name{NAFOdown-package}
\alias{NAFOdown}
\alias{NAFOdown-package}
\title{NAFOdown: An R Markdown Template for Producing NAFO Scientific Council Documents}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

NAFOdown uses the bookdown package to create NAFO documents, such as SCRs or STACFIS reports, in Word format. The goal of this package is to encourage reproducible research using simple Markdown syntax and embedded R code, and facilitate the process of producing or updating scientific documents that consistently follow NAFO formatting guidelines. This package is based on Sean Anderson's csasdown and Chester Ismay's thesisdown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nafc-assess/NAFOdown}
  \item Report bugs at \url{https://github.com/nafc-assess/NAFOdown/issues}
}

}
\author{
\strong{Maintainer}: Paul Regular \email{Paul.Regular@dfo-mpo.gc.ca}

Authors:
\itemize{
  \item Bob Rogers
}

Other contributors:
\itemize{
  \item Laura Wheeland [contributor]
  \item Andrea Perreault [contributor]
  \item Mariano Koen-Alonso [contributor]
  \item Sean Anderson (Development of csasdown) [contributor]
  \item Chris Grandin (Development of csasdown) [contributor]
  \item Chester Ismay (Development of thesisdown) [contributor]
  \item Nick Solomon (Development of thesisdown) [contributor]
}

}
\keyword{internal}
