% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_lengths}
\alias{plot_lengths}
\title{Generate ridges plot from annual length frequency data}
\usage{
plot_lengths(
  data,
  rows = NULL,
  cols = NULL,
  xlab = "Length (cm)",
  ylab = "Year",
  scale = 5,
  alpha = 0.8,
  scale_color = scale_color_viridis_d,
  scale_fill = scale_fill_viridis_d,
  drop_zeros = TRUE
)
}
\arguments{
\item{data}{A data.frame with \code{"year"}, \code{"length"}, and \code{"frequency"} headers. Extra columns
may aso be utilized to define facets using the rows and cols arguments.}

\item{rows}{Column to use to define facet rows.}

\item{cols}{Column to use to define facet columns.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{scale}{Scaling factor controlling the height of the ridges.}

\item{alpha}{Value controling the transparancy of the ridges.}

\item{scale_color}{Function such as \link[ggplot2:scale_viridis]{ggplot2::scale_colour_viridis_d} for defining line
colours to use across years.}

\item{scale_fill}{Function such as \link[ggplot2:scale_viridis]{ggplot2::scale_fill_viridis_d} for defining fill
colours to use across years.}

\item{drop_zeros}{Should zeros be replaced by NA values? (i.e., impose blank space in ridge plot).}
}
\value{
Returns a ggplot2 object that can be modified further by functions such as \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
Function for producing a faceted ridge plot of annual length frequencies using the ggridges package.
}
\examples{

## Simulate some data
set.seed(3)
samples_per_year <- 1000
years <- 2000:2022
lambdas <- rnorm(length(years), mean = 25, sd = 10)
data <- lapply(seq_along(years), function(i) {
    lengths <- rpois(samples_per_year, lambda = lambdas[i])
    lf <- table(lengths)
    data.frame(year = years[i], length = as.numeric(names(lf)), frequency = as.numeric(unname(lf)))
})
data <- do.call(rbind, data)

## Using defaults
plot_lengths(data)

## Customize using ggplot functions
plot_lengths(data, alpha = 0.1, scale = 2) +
scale_color_manual(values = rep("grey30", length(unique(data$year)))) +
scale_fill_manual(values = rep("grey30", length(unique(data$year))))


}
