% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{scale_color_nafo}
\alias{scale_color_nafo}
\alias{scale_colour_nafo}
\alias{scale_fill_nafo}
\title{NAFO color palette for ggplot2}
\usage{
scale_color_nafo(...)

scale_colour_nafo(...)

scale_fill_nafo(...)
}
\arguments{
\item{...}{Additional arguments passed to the scale functions.}
}
\value{
A ggplot2 scale function.
}
\description{
These functions provide custom color and fill scales for ggplot2 plots, based on a NAFO color palette.
}
