% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{theme_nafo}
\alias{theme_nafo}
\title{ggplot2 theme for styling plots for NAFO documents}
\usage{
theme_nafo(base_size = 9, base_family = "Cambria")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
This theme will produce ggplots that follow NAFO plot guidelines:
\itemize{
\item Font: Cambria 9pt.
\item Plot Border: black 0.5 pt. No borders around entire graph, just plot. Borders around legend (0.5pt) if possible.
\item Tick Marks to the inside on x & y axis
\item Graph lines: 0.75 pt, black if possible, otherwise grey
\item Plot size: Height 6.4 cm, width 11.5 cm.
}

The helper values \link{.nafo_lwd}, \link{.nafo_height}, and \link{.nafo_width} can be used to impose consistent graph line widths, plot heights, and plot widths, respectively.
}
\examples{

library(ggplot2)
set.seed(123)
d <- data.frame(SSB = rlnorm(40), Recruitment = rlnorm(40),
                Species = c(rep("Cod", 20), rep("Plaice", 20)))
ggplot(d) +
geom_point(aes(x = SSB, y = Recruitment, shape = Species),
               fill = "white", color = "black", shape = 21,
               size = .nafo_pts, stroke = .nafo_stroke) +
geom_hline(aes(yintercept = mean(Recruitment)),
           linetype = 2, size = .nafo_lwd) +
theme_nafo()

}
\seealso{
\link[ggplot2:ggtheme]{ggplot2::theme_minimal}
}
