% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{theme_nafotabs}
\alias{theme_nafotabs}
\title{flextable theme for styling tables for NAFO docuemtns}
\usage{
theme_nafotabs(x, fontsize = 11, fontname = "Cambria")
}
\arguments{
\item{x}{A flextable object}

\item{fontsize}{Font size}

\item{fontname}{Font family}
}
\description{
A modification of \link[flextable:theme_booktabs]{flextable::theme_booktabs}
}
\examples{

library(flextable)
d <- data.frame(SSB = rlnorm(10), Recruitment = rlnorm(10))
flextable(d) \%>\%
  theme_nafotabs()
}
