% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_stock}
\alias{toy_stock}
\title{Toy Stock Data}
\format{
A data frame with 40 rows and 13 variables:
\describe{
\item{year}{Year of the simulation, ranging from 1981 to 2020.}
\item{biomass}{Simulated biomass of the fish stock.}
\item{rec_index1}{Simulated recruitment index 1.}
\item{rec_index2}{Simulated recruitment index 2.}
\item{rec_index3}{Simulated recruitment index 3.}
\item{TAC}{Total allowable catch (TAC) for the fish stock.}
\item{catch}{Simulated catch of the fish stock.}
\item{Fmsy}{Fishing mortality rate at maximum sustainable yield.}
\item{Bmsy}{Biomass at maximum sustainable yield.}
\item{Brel}{Relative biomass compared to 30\% of Bmsy.}
\item{Brel_lwr}{Lower bound of relative biomass confidence interval.}
\item{Brel_upr}{Upper bound of relative biomass confidence interval.}
\item{Frel}{Relative fishing mortality rate compared to Fmsy.}
\item{Frel_lwr}{Lower bound of relative fishing mortality rate confidence interval.}
\item{Frel_upr}{Upper bound of relative fishing mortality rate confidence interval.}
}
}
\usage{
toy_stock
}
\description{
A simulated dataset for demonstrating plotting functions in the NAFOdown R package. The dataset contains simulated fish stock dynamics over a 40-year period.
}
\details{
The data were generated using a simple surplus production model.
}
\keyword{datasets}
