% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{nafo_summary_table}
\alias{nafo_summary_table}
\title{Function for producing summary tables for NAFO documents}
\usage{
nafo_summary_table(comments = rep(NA, 5), status = rep(NA, 5))
}
\arguments{
\item{comments}{A vector of length 5 elaborating on the status the status of each 'Convention General Principle'}

\item{status}{A vector of length 5 decribing the status of each 'Convention General Principle'. Must be one of: 'OK', 'Intermediate','Not accomplished', 'Unknown'}
}
\description{
Produces summary tables for NAFO stock summary sheets
}
