% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{theme_nafo}
\alias{theme_nafo}
\title{NAFO plotting theme for ggplot2}
\usage{
theme_nafo(base_size = 9, base_family = "Cambria")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
This theme will produce ggplots that follow NAFO plot guidelines:
\itemize{
\item Font: Cambria 9pt.
\item Plot Border: black 0.5 pt. No borders around entire graph, just plot. Borders around legend (0.5pt) if possible.
\item Tick Marks to the outside on x & y axis
\item Graph lines: 0.75 pt, black if possible, otherwise grey
\item Plot size: Height 6.4 cm, width 11.5 cm.
}

The helper values \link{.nafo_lwd}, \link{.nafo_height}, and \link{.nafo_width} can be used to impose consistent graph line widths, plot heights, and plot widths, respectively.
}
\examples{

library(ggplot2)
library(patchwork)
showtext::showtext_opts(dpi = 96) # Adjust dpi to match your plotting device (e.g., set to 300 if using ggsave with dpi set to 300)
theme_set(theme_nafo()) # set default ggplot2 theme

C_plot <- ggplot(aes(x = year), data = toy_stock) +
    geom_bar(aes(y = catch, fill = "Catch"), stat = "identity", alpha = 0.4) +
    geom_step(aes(x = year - 0.5, y = ifelse(year > 2015, 0, TAC), linetype = "TAC"), color = .nafo_cols[2]) +
    scale_y_continuous(n.breaks = 10, name = "Catch / TAC", expand = expansion(c(0.01, 0.05))) +
    scale_x_continuous(n.breaks = 10, name = "", expand = expansion(c(0.01, 0.01))) +
    scale_fill_nafo() +
    theme(legend.title = element_blank(),
          legend.spacing.y = unit(0.1, "cm"))

B_plot <- ggplot(aes(x = year), data = toy_stock) +
    geom_ribbon(aes(ymin = relB_lwr, ymax = relB_upr), fill = .nafo_cols[1], alpha = 0.4) +
    geom_line(aes(y = relB), color = .nafo_cols[1]) +
    scale_y_continuous(limits = c(0, NA), n.breaks = 10, name = "B/Blim", expand = expansion(c(0.01, 0.05))) +
    scale_x_continuous(n.breaks = 10, name = "", expand = expansion(c(0.01, 0.01)))

F_plot <- ggplot(aes(x = year), data = toy_stock) +
    geom_ribbon(aes(ymin = relF_lwr, ymax = relF_upr), fill = .nafo_cols[1], alpha = 0.4) +
    geom_line(aes(y = relF), color = .nafo_cols[1]) +
    scale_y_continuous(limits = c(0, NA), n.breaks = 10, name = "F/Flim", expand = expansion(c(0.01, 0.05))) +
    scale_x_continuous(n.breaks = 10, name = "", expand = expansion(c(0.01, 0.01)))

R_plot <- ggplot(aes(x = year), data = toy_stock) +
    geom_line(aes(y = rec_index1, color = "Survey 1")) +
    geom_line(aes(y = rec_index2, color = "Survey 2")) +
    geom_line(aes(y = rec_index3, color = "Survey 3")) +
    scale_y_continuous(limits = c(0, NA), name = "Recruitment index", expand = expansion(c(0.01, 0.05))) +
    scale_x_continuous(n.breaks = 10, name = "", expand = expansion(c(0.01, 0.01))) +
    scale_color_nafo() +
    theme(legend.title = element_blank())

(C_plot + B_plot) / (F_plot + R_plot)

if (interactive()) {
    showtext::showtext_opts(dpi = 300)
    ggsave("summary_plot.png", height = .nafo_height * 2, width = .nafo_width * 2)
}

}
\seealso{
\link[ggplot2:ggtheme]{ggplot2::theme_minimal}
}
