% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{theme_NAFO}
\alias{theme_NAFO}
\title{ggplot2 theme for styling plots for NAFO docuemtns}
\usage{
theme_NAFO(base_size = 14, base_family = "serif")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
ggplot2 theme for styling plots for NAFO docuemtns
}
\examples{

d <- data.frame(SSB = rlnorm(20), Recruitment = rlnorm(20))
ggplot(d) +
geom_point(aes(x = SSB, y = Recruitment)) +
theme_NAFO()

}
\seealso{
\link[ggplot2:theme_minimal]{ggplot2::theme_minimal}
}
