% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{word_scr}
\alias{word_scr}
\alias{word_stacfis}
\title{Creates an R Markdown Word NAFO-formatted document}
\usage{
word_scr(...)

word_stacfis(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link[bookdown:word_document2]{bookdown::word_document2()}}}
}
\value{
A Word Document based on the NAFO SCR word template.
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of an SCR or a
STACFIC report.
}
