% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{word_scr}
\alias{word_scr}
\alias{word_stacfis}
\alias{word_scs}
\title{Creates an R Markdown Word NAFO-formatted document}
\usage{
word_scr(base_format = officedown::rdocx_document, ...)

word_stacfis(base_format = officedown::rdocx_document, ...)

word_scs(base_format = officedown::rdocx_document, ...)
}
\arguments{
\item{...}{arguments to pass to \code{\link[bookdown:word_document2]{bookdown::word_document2()}}}
}
\value{
A Word Document based on the NAFO SCR, STACFIS or SCS word template.
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of an SCR, STACFIC
or SCS report.
}
